+++ src/components/Deriv/Deriv.js
@@ -6,6 +6,7 @@ import PropTypes from 'prop-types'
 import styles from './styles'
 import Notice from '../Notice/Notice'
 import ButtonLink from '../Button/ButtonLink'
+import DerivForm from './DerivForm'
 
 import { canonical, siteTitle } from '../../App'
 
@@ -22,6 +23,7 @@ const Deriv = props => (
     <ButtonLink route='/deriv/help'>Help</ButtonLink>
     <h2 className='toolTitle'>LanguaDeriv</h2>
     <Notice>This tool is still in planning.</Notice>
+    <DerivForm styles={props.classes} />
   </div>
 )