+++ src/components/Tree/Tree.js
@@ -6,6 +6,7 @@ import PropTypes from 'prop-types'
 import styles from './styles'
 import Notice from '../Notice/Notice'
 import ButtonLink from '../Button/ButtonLink'
+import TreeForm from './TreeForm'
 
 import { canonical, siteTitle } from '../../App'
 
@@ -22,6 +23,7 @@ const Tree = props => (
     <ButtonLink route='/tree/help'>Help</ButtonLink>
     <h2 className='toolTitle'>LanguaTree</h2>
     <Notice>This tool is still in planning.</Notice>
+    <TreeForm styles={props.classes} />
   </div>
 )